fix: user.ts 타입 분리 → 분석영상, 크레딧, 결제 타입
