+++ types/credit.ts
@@ -1,0 +1,6 @@
+export interface CreditRecord {
+  id: string;
+  date: string;
+  change: number;
+  type: 'charge' | 'use';
+}
