+++ types/purchase.ts
@@ -1,0 +1,6 @@
+export interface PurchaseRecord {
+  id: string;
+  plan: string;
+  price: number;
+  date: string;
+}
